==> notebooks/Step1_NLP.ipynb <==
# Nạp các gói thư viện NumPy và Pandas
import pandas as pd
import numpy as np
import regex
import string

#Tạo đường link
import sys
import os

current = os.path.abspath('')
parent = os.path.dirname(current)
grandparent = os.path.dirname(parent)

#File dữ liệu

#File data media
title_file = parent+"/data/processed/media07_01_2023.xlsx"

#Read file
title = pd.read_excel(parent+"/data/processed/media07_01_2023.xlsx")

# Underthesea là module mã nguồn mở được dùng để hỗ trợ cho việc phân tích tiếng Việt.
# Nguồn: https://github.com/undertheseanlp/underthesea
from underthesea import word_tokenize, pos_tag, sent_tokenize

# TextProcessor là gói thư viện hỗ trợ HV môn MDS5 (của trung tâm tin học Đại học Khoa học Tự Nhiên) 
# thực hiện các bước xử lý 1 số vấn đề thường gặp ở văn bản Tiếng Việt.
# Nguồn: https://pypi.org/project/ttth-mds5-analyzer/
from processor.text import TextProcessor
text_processor = TextProcessor()

def remove_stockword(text, stockword):
    ###### REMOVE stock words
    document = ' '.join('' if word.upper() in stockword else word for word in text.split())
    ###### DEL excess blank space
    document = regex.sub(r'\s+', ' ', document).strip()
    return document

def remove_time(text):
    ###### REMOVE time format
    document = ' '.join('' if word.find('/')!=-1 else word for word in text.split())
    ###### DEL excess blank space
    document = regex.sub(r'\s+', ' ', document).strip()
    return document

def remove_timeword(text):
    ###### REMOVE time words
    document = ' '.join('' if word in ['ngày', 'tháng', 'năm', 'quý', 'lần'] else word for word in text.split())
    ###### DEL excess blank space
    document = regex.sub(r'\s+', ' ', document).strip()
    return document

def clean_text(text, stockword_list):
    # Chuyển tất cả chữ trong câu thành chữ thường
    sentence = text.lower()
    
    # Loại bỏ các ký tự số hoặc dấu câu
    sentence = text_processor.remove_punctuation_number(sentence)
    
    # Chuẩn hoá sang unicode
    sentence = text_processor.covert_unicode(sentence)
    
    # Xử lý từ ghép
    sentence = text_processor.process_postag_thesea(sentence,
                                                    lst_word_type = ['A', 'AB', 'V', 'VB', 'VY', 'R', 'N'])
    
    # Xử lý từ có không
    sentence = text_processor.process_special_word(sentence)
    
    sentence = remove_stockword(sentence, stockword_list)
    sentence = remove_timeword(sentence)
    sentence = remove_time(sentence)
    return sentence

clean_text('Nghị quyết HĐQT v/v triệu tập ĐHĐCĐ bất thường năm 2023', ['ABC'])

# Nap du lieu media

title = title[['Date', 'Time', 'Ma_CK', 'Title']]
title.head(10)

stockword = title['Ma_CK'].unique().tolist()
print(stockword)

title['Time'] = title['Time'].str[:4].str.replace(':', '.').astype('float')
title['Date'] = pd.DatetimeIndex(title['Date'])
title = title.rename(columns={'Date':'date',
                              'Time':'hour',
                              'Ma_CK':'ma_ck',
                              'Title':'title'})
title = title.dropna().reset_index(drop=True)
title.head(10)

# Xu ly NLP

full_content_new = []
for i in range(len(title)):
    text = title.loc[i]["title"]
    full_content_new.append(clean_text(text, stockword))

processed_data = pd.DataFrame.from_dict({"title":title["title"], "title_new":full_content_new})
processed_data.head(10)

title_backup = title.copy()

from datetime import datetime, timedelta
title = title_backup.copy()
title['weekday'] = title['date'].dt.day_name()
title['date'][title['weekday']=='Sunday'] = title[title['weekday']=='Sunday']['date'] - timedelta(days=2)
title['date'][title['weekday']=='Saturday'] = title[title['weekday']=='Saturday']['date'] - timedelta(days=1)

title['weekday'].value_counts()

title['hour'][title['weekday']=='Sunday'] = 19
title['hour'][title['weekday']=='Saturday'] = 19
title

# Ghep bang gia va NLP

df_delta = title.copy()
df_delta['title_new'] = processed_data['title_new']
df_delta.head(10)

#Save file
df_delta.to_csv(parent+'/data/processed/NLP.csv', index=False)
