==> notebooks/Step2_Calculate_stock_return_and_ volatility.ipynb <==
# Nạp các gói thư viện NumPy và Pandas
import pandas as pd
import numpy as np
import regex
import string

#Tạo đường link
import sys
import os

current = os.path.abspath('')
parent = os.path.dirname(current)
grandparent = os.path.dirname(parent)

#File path

## File media data
title_file = parent+"/data/processed/NLP.csv"

## File stock data
price_file = parent+'/data/processed/price07_01_2023.xlsx'

# Nap du lieu media

title = pd.read_csv(title_file)
title['date'] = pd.to_datetime(title['date'])
title

# Nap du lieu gia chung khoan

price_root = pd.read_excel(price_file, usecols = ['Date', 'maCK', 'Open', 'Close'])

price = price_root.copy()
price['Date'] = pd.DatetimeIndex(price['Date'], dayfirst=True)
price = price.rename(columns={'Date':'date',
                              'maCK':'ma_ck',
                              'Open':'open_price',
                              'Close':'close_price'})
price = price.sort_values(['ma_ck', 'date'], ascending=True)
price['open_price'] = price['open_price'].astype('str')
price['open_price'] = price['open_price'].str.replace('00.0','000')
price['open_price'] = price['open_price'].str.replace(',', '.')
# price = price[price['open_price'].isnull()==False]
price

# Tam bo PVX va PXT
price['open_price'] = price['open_price'].astype('float', errors='raise')
price = price.drop_duplicates(ignore_index=True)
price

# Xu ly du lieu chung khoan

price['close_price_lag'] = price.groupby('ma_ck')['close_price'].shift()
price['open_price_next'] = price.groupby('ma_ck')['open_price'].shift(-1)
price = price.dropna()
price

price['return_in_session'] = np.log(price['close_price']/price['open_price'])
price['return_before_session'] = np.log(price['open_price']/price['close_price_lag'])
price

price_stack = price[['date', 'ma_ck', 'return_in_session', 'return_before_session']].set_index(['date', 'ma_ck']).stack().reset_index()
price_stack.columns = ['date', 'ma_ck', 'session', 'return']
price_stack = price_stack.sort_values(['ma_ck', 'date', 'session'])
price_stack = price_stack[price_stack['return']!=np.inf]
price_stack = price_stack[price_stack['return']!=-np.inf]

# Ghep bang gia va NLP

from datetime import datetime, timedelta
title['new_date'] = title['date']
title['weekend'] = False
title['weekend'][title['weekday']=='Friday'] = title[title['weekday']=='Friday']['hour']>16
title['new_date'][title['hour']>16] = title[title['hour']>16]['new_date']+timedelta(days=1)
title['new_date'][title['weekend']==True] = title[title['weekend']==True]['new_date']+timedelta(days=2)
title['session'] = ['return_in_session' if (x>8) and (x<16) else 'return_before_session' for x in title['hour']]
title

price_stack = price_stack.rename(columns={'date':'new_date'})
df_delta = title.merge(price_stack, on=['new_date', 'ma_ck', 'session'], how='left')
df_delta

df_delta = df_delta.drop_duplicates(ignore_index=True)
df_delta = df_delta.dropna().reset_index(drop=True)
df_delta                                    

df_delta_sub = df_delta
df_delta_sub['title_new'] = df_delta_sub['title_new']+' '
df_final = df_delta_sub[['ma_ck', 'title_new', 'new_date', 'session', 'return']].groupby(['new_date', 'ma_ck', 'session', 'return']).sum().reset_index()
df_final['title_new'] = df_final['title_new'].str[:-1]
df_final

#Save file
df_final.to_csv(parent+'/data/processed/input_table.csv', index=False)
