==> notebooks/Step3_Vectorizer_and_Cleaning.ipynb <==
# Nạp các gói thư viện NumPy và Pandas
import pandas as pd
import numpy as np
import regex
import string

#Tạo đường link
import sys
import os

current = os.path.abspath('')
parent = os.path.dirname(current)
grandparent = os.path.dirname(parent)

#File path
df_file = parent+"/data/processed/input_table.csv"

df = pd.read_csv(df_file)
df = df.dropna().reset_index(drop=True)
df

# Loai bo nhung dong return la 0
df = df[df['return']!=0]

# Tao bien deep de giu lai nhung dong return thuc su co y nghia
df['deep']=0
df['deep'][df['return']>df['return'].quantile(0.999)]=100
df['deep'][df['return']<df['return'].quantile(0.001)]=100
df = df.reset_index(drop=True)

#Tao bien dummies ma chung khoan
dummies = pd.get_dummies(df['ma_ck'])
dummies

# Vectorize

# bag of words
from sklearn.feature_extraction.text import CountVectorizer

X = df['title_new']
y = df['return']

vectorizer = CountVectorizer()
vectorizer.fit(X)
X_vec = vectorizer.transform(X)
X_vec

tfidf_feature_names = vectorizer.get_feature_names_out()
tfidf_feature_names

# Tao bang sentiment gom nhung tu da duoc vectorize
sentiments = pd.DataFrame(data=X_vec.toarray(), columns = tfidf_feature_names)
sentiments[dummies.columns] = dummies

sentiments['deep'] = df['deep']

#Nhung tu ngu co anh huong lon
deep = pd.DataFrame(data = sentiments[sentiments['deep']==100].sum(), columns=['deep'])
deep = deep[deep['deep']>0]
deep_lst = deep.index.tolist()
deep_lst

# Top những từ xuất hiện nhiều hơn 1000 lần
big = pd.DataFrame(data=np.count_nonzero(sentiments, axis=0), index=sentiments.columns, columns=['counts'])
big_lst = big[big['counts']>0.01*len(sentiments)].index.tolist()
big_lst

for i in deep_lst:
  if i not in big_lst:
    big_lst.append(i)

big_lst

sentiments_new = sentiments[big_lst]
sentiments_new['title'] = df['title_new']
sentiments_new['return'] = df['return']
sentiments_new['ma_ck'] = df['ma_ck']
sentiments_new['Date'] = df['new_date']
sentiments_new['session'] = df['session']
sentiments_new['sum'] = sentiments_new[big_lst].sum(axis=1)
sentiments_new = sentiments_new[sentiments_new['sum']>sentiments_new['sum'].quantile(0.2)]
sentiments_new

for i in sentiments_new.columns:
  if '.1' in i:
    print(i)

# Save file
sentiments_new.to_csv(parent+'/data/processed/sentiments_vector.csv', index=False)
