==> notebooks/Step4_Modeling_finance_dictionary_Sklearn_Classification.ipynb <==
# Nạp các gói thư viện NumPy và Pandas
import pandas as pd
import numpy as np
import regex
import string

#Tạo đường link
import sys
import os

current = os.path.abspath('')
parent = os.path.dirname(current)
grandparent = os.path.dirname(parent)

#File path

#Sentiment vector
df_file = parent+"/data/processed/sentiments_vector.csv"

#Media data
title_file = parent+"/data/processed/media07_01_2023.xlsx"

df = pd.read_csv(df_file)
df

# Tao bien binary cho return
def get_binary(x):
  if x>=0:
    return 0
  if x<0:
    return 1

df['return'] = df['return'].apply(get_binary)
df

# Rút ra danh sách các từ khoá

sentiments = df.columns.tolist()[:]
for i in ['title', 'return', 'deep', 'ma_ck', 'Date', 'sum', 'session']:
  if i in sentiments:
    sentiments.remove(i)

X = df[sentiments]
y = df['return'].astype('int')

# MODELLING

# Scaling X
# from sklearn.preprocessing import StandardScaler
# sentiments_scaled = pd.DataFrame(data = StandardScaler().fit_transform(W))
sentiments_scaled = X

# Split train and test sample
# from sklearn.model_selection import train_test_split
# X_train, X_test, y_train, y_test = train_test_split(sentiments_scaled, y, test_size=0.2, random_state=42)
X_train = X
X_test = X
y_train = y
y_test = y

X_train

y_train.shape

#FITTING THE CLASSIFICATION MODEL using Random Forest
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import classification_report, f1_score, accuracy_score, confusion_matrix,roc_curve,auc
model=RandomForestClassifier(random_state = 42)
model.fit(X_train, y_train.values)

#Predict y value for test dataset
y_pred = model.predict(X_test)
y_prob = model.predict_proba(X_test)[:,1]


print(classification_report(y_test,y_pred))
print('Confusion Matrix:',confusion_matrix(y_test, y_pred))

fpr, tpr, thresholds = roc_curve(y_test, y_prob)
roc_auc = auc(fpr, tpr)
print('AUC:', roc_auc)

result = df[['Date', 'ma_ck', 'title']]
result['negative'] = model.predict(df[sentiments])
result['negative_prob'] = model.predict_proba(df[sentiments])[:,1]
result = result.rename(columns={'title':'title_new'})

result

# Save model
import pickle
filename = parent+'/data/processed/classification_model.sav'
pickle.dump(model, open(filename, 'wb'))
